Conflicts:
	astrid/AndroidManifest.xml
	astrid/plugin-src/com/todoroo/astrid/producteev/sync/ProducteevDataService.java
	astrid/plugin-src/com/todoroo/astrid/rmilk/MilkBackgroundService.java
	astrid/plugin-src/com/todoroo/astrid/rmilk/MilkLoginActivity.java
	astrid/plugin-src/com/todoroo/astrid/rmilk/data/MilkDataService.java
	astrid/plugin-src/com/todoroo/astrid/rmilk/sync/RTMSyncProvider.java
	astrid/src/com/todoroo/astrid/service/StartupService.java

+++ astrid/src/com/todoroo/astrid/activity/TaskEditActivity.java
@@ -104,12 +104,17 @@ public final class TaskEditActivity extends TabActivity {
     /**
      * Task ID
      */
-    public static final String TOKEN_ID = "i"; //$NON-NLS-1$
+    public static final String TOKEN_ID = "id"; //$NON-NLS-1$
 
     /**
      * Content Values to set
      */
     public static final String TOKEN_VALUES = "v"; //$NON-NLS-1$
+
+    /**
+     * Task in progress (during orientation change)
+     */
+    private static final String TASK_IN_PROGRESS = "task_in_progress"; //$NON-NLS-1$
 
     // --- request codes
 
@@ -185,6 +190,16 @@ public final class TaskEditActivity extends TabActivity {
 
 		// disable keyboard until user requests it
 		AndroidUtilities.suppressVirtualKeyboard(title);
+
+		// if we were editing a task already, restore it
+		if(savedInstanceState != null && savedInstanceState.containsKey(TASK_IN_PROGRESS)) {
+		    Task task = savedInstanceState.getParcelable(TASK_IN_PROGRESS);
+		    if(task != null) {
+		        model = task;
+		    }
+		}
+
+		setResult(RESULT_OK);
     }
 
     /* ======================================================================
@@ -304,6 +319,12 @@ public final class TaskEditActivity extends TabActivity {
      */
     @SuppressWarnings("nls")
     protected void loadItem(Intent intent) {
+        if(model != null) {
+            // came from bundle
+            isNewTask = (model.getValue(Task.TITLE).length() == 0);
+            return;
+        }
+
         long idParam = intent.getLongExtra(TOKEN_ID, -1L);
 
         database.openForReading();
@@ -323,6 +344,9 @@ public final class TaskEditActivity extends TabActivity {
         if(model.getValue(Task.TITLE).length() == 0) {
             FlurryAgent.onEvent("create-task");
             isNewTask = true;
+
+            // set deletion date until task gets a title
+            model.setValue(Task.DELETION_DATE, DateUtilities.now());
         } else {
             FlurryAgent.onEvent("edit-task");
         }
@@ -354,7 +378,10 @@ public final class TaskEditActivity extends TabActivity {
         for(TaskEditControlSet controlSet : controls)
             controlSet.writeToModel(model);
 
-        if(taskService.save(model, false))
+        if(title.getText().length() > 0)
+            model.setValue(Task.DELETION_DATE, 0L);
+
+        if(taskService.save(model) && title.getText().length() > 0)
             showSaveToast();
     }
 
@@ -362,11 +389,11 @@ public final class TaskEditActivity extends TabActivity {
     public void finish() {
         super.finish();
 
-        // abandon editing in this case
-        if(title.getText().length() == 0 && isNewTask) {
+        // abandon editing and delete the newly created task if
+        // no title was entered
+
+        if(title.getText().length() == 0 && isNewTask && model.isSaved()) {
             taskService.delete(model);
-            showCancelToast();
-            setResult(RESULT_CANCELED);
         }
     }
 
@@ -418,10 +445,6 @@ public final class TaskEditActivity extends TabActivity {
      * precision
      */
     private void showSaveToast() {
-        // if we have no title, or nothing's changed, don't show toast
-        if(isNewTask)
-            return;
-
         int stringResource;
 
         long due = model.getValue(Task.DUE_DATE);
@@ -492,7 +515,6 @@ public final class TaskEditActivity extends TabActivity {
                 Toast.LENGTH_SHORT).show();
     }
 
-
     @Override
     public boolean onMenuItemSelected(int featureId, MenuItem item) {
         switch(item.getItemId()) {
@@ -532,10 +554,11 @@ public final class TaskEditActivity extends TabActivity {
 
     @Override
     protected void onPause() {
-        if(shouldSaveState)
-            save();
         super.onPause();
         unregisterReceiver(controlReceiver);
+
+        if(shouldSaveState)
+            save();
     }
 
     @Override
@@ -544,6 +567,14 @@ public final class TaskEditActivity extends TabActivity {
         registerReceiver(controlReceiver,
                 new IntentFilter(AstridApiConstants.BROADCAST_SEND_EDIT_CONTROLS));
         populateFields();
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+
+        // stick our task into the outState
+        outState.putParcelable(TASK_IN_PROGRESS, model);
     }
 
     @Override
